// brokerchain-1.0.0/src/library/file-system/file-rename/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    base_path: string;
    name_from: string;
    name_to: string;
}

export interface OutputOk {}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "base_path") return;
            if (field === "name_from") return;
            if (field === "name_to") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.base_path must be string");
        if (typeof v.base_path !== "string") {
            throw new Error("v.base_path is not string");
        }

        log.println("v.name_from must be string");
        if (typeof v.name_from !== "string") {
            throw new Error("v.name_from is not string");
        }

        log.println("v.name_to must be string");
        if (typeof v.name_to !== "string") {
            throw new Error("v.name_to is not string");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {};
        return obj;
    } else {
        return {};
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;